Add basic request scheduling and time recording (#4)
* Add functionality to simulate scheduling requests

Add a Runner which serves as the manager of the simulation. Also
add scheduler which manages scheduling block requests. Add algorithms
which form the basis for scheduling according to different algorithms.

Change DiskSimulation to run a simple FIFO algorithm for demo
purposes. Add variable currentCylinderPosition to Disk class.

* Change method 'getCurrentCylinderPosition' name

Change name of method to 'getCurrentHeadPosition'

* Change default variables to final

Change variables that are 'defaults' to public final. Ensures other
classes can read these parameters and not change them.