Time hotfix - Closes #5 (#6)
* Fix runner loop bug

The method run() in runner would loop through on only the request list
and not the queue. This would make run stop processing the queue once
the request list is empty, which is not intended functionality.

Change the loop to iterate until the queue list is empty. Add helper
method to Scheduler to assist with returning queue. Modify
Scheduler's fillQueue() method to only fill queue if requests are
present, which assists with fixing the loop iteration bug.

Add toString() method to Request which pretty prints a requests
attributes to assist with debugging.

* Fix time bug

Add time each request seeks to each request in the queue. This
makes the totalQueueTime metric function as intended.

Add printout of metrics of total queue time and average queue wait
time to main(). This makes the program's output more understandable.

* Fix #5

Fixes total time calculation. This ensures the program is functioning
according the the assignment specification.

Update comments for understandability.