Add logic for SSTF algorithm (#7)
Add logic to SSTF class to simulate the SSTF algorithm as specified
by the assignment specification.

Modified Algorithm abstract class to include the current head position
as a parameter for the sort() method. This allows certain algorithms,
like SSTF, to function properly since they require knowing the current
position of the disk's head.

Modified the rounding value for seek time in Disk to round to two
decimal places in order to make the average calculation more precise.