Add CSCAN algorithm and update main method (#8)
* Begin Cscan implementation

Add currentHeadMovement variable to abstract algorithm class in order
to keep track of the direction the disk head is moving.

Remove Scan class from project since the Scan algorithm is not part
of the assignment specificiation.

* Add logic for CSCAN algorithm

Add ability to simulate CSCAN algorithm as required by the assignment
specification.

Note that the implemented CSCAN algorithm is a crude, non efficient
way to go about simulating CSCAN due to the many sorts called upon the
request queue. While there is most likely a more efficient way to
implement the CSCAN algorithm, this is just a proof of concept.

* Remove head movement variable

Remove head movement variable from Algorithm and Disk since it is not
needed.

* Update formatting and comments

Update comments, code organization, and formatting for readability
of program.

* Remove print statement from Scheduler

* Updated formatting and readability of code

* Add logic to run simulation based on given args

Change main() to use given arguments to create a new simulation based
off of those arguments according to the assignment specification.

Add print statement to inform user what algorithm is running along
with queue depth and file location.

Add getName() method to Algorithm to be able to retrieve the name for
a given algorithm.